+++ NEWS.md
@@ -1,4 +1,21 @@
 
+
+
+# Ramses 0.3.0
+
+## Features
+
+* introduce objects
+
+## Improvements
+
+* remove dependencies
+* improve handling of time data by SQLite
+
+## Documentation
+
+* new ramses-objects vignette
+
 # Ramses 0.2.0
 
 * Support PostgreSQL